% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_states_panel.R
\name{build_states_cow_panel}
\alias{build_states_cow_panel}
\title{Build a state panel}
\usage{
build_states_cow_panel(
  start_year = 1946,
  end_year = 2013,
  exclude_microstates = TRUE,
  exclude_non_un = TRUE,
  exclude_islands = FALSE
)
}
\arguments{
\item{start_year}{First year.}

\item{end_year}{Last year.}

\item{exclude_microstates}{Logical.}

\item{exclude_non_un}{Logical.}

\item{exclude_islands}{Logical.}

\item{coding_system}{Country coding system. Either \code{"cow"} or \code{"gw"}.}
}
\value{
A data frame with state-year observations.
}
\description{
Build a state panel
}
