% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_year_range}
\alias{check_year_range}
\title{Check year range}
\usage{
check_year_range(start_year, end_year)
}
\arguments{
\item{start_year}{First year.}

\item{end_year}{Last year.}
}
\description{
Check year range
}
\keyword{internal}
