% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_coding_system}
\alias{detect_coding_system}
\title{Detect coding system from a panel data frame}
\usage{
detect_coding_system(panel)
}
\arguments{
\item{panel}{A data frame produced by \code{build_states_panel()}.}
}
\value{
\code{"cow"} or \code{"gw"}.
}
\description{
Detect coding system from a panel data frame
}
\keyword{internal}
