% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vdem_data.R
\name{load_vdem_data}
\alias{load_vdem_data}
\title{Load V-Dem data}
\source{
Varieties of Democracy (V-Dem) Project. Please cite V-Dem when using
these data.
}
\usage{
load_vdem_data(
  vars = NULL,
  start_year = 1945,
  end_year = 2019,
  coding_system = c("cow", "gw")
)
}
\arguments{
\item{vars}{Character vector of V-Dem variable names to include. If \code{NULL},
a default set of civil society, repression, civil liberties, democracy,
corruption, and rule-of-law indicators is returned.}

\item{start_year}{Integer. First year to include.}

\item{end_year}{Integer. Last year to include.}

\item{coding_system}{Character. Country coding system to use.
Either \code{"cow"} for Correlates of War codes or \code{"gw"} for
Gleditsch-Ward codes.}
}
\value{
A data frame with country-year V-Dem indicators. If
\code{coding_system = "cow"}, the data frame contains a \code{cow} column.
If \code{coding_system = "gw"}, it contains a \code{gw} column.
}
\description{
Loads selected country-year variables from the V-Dem dataset provided by
the \code{vdemdata} package. The returned data can use either COW or
Gleditsch-Ward country codes.
}
\details{
This function requires the \code{vdemdata} package. If it is not installed,
install it with:

\code{remotes::install_github("vdeminstitute/vdemdata")}

For \code{coding_system = "gw"}, V-Dem COW codes are converted to
Gleditsch-Ward codes using \code{countrycode::countrycode()}.
}
\examples{
\dontrun{
vdem_cow <- load_vdem_data(coding_system = "cow")

vdem_selected <- load_vdem_data(
  vars = c("v2x_polyarchy", "v2x_libdem", "v2x_rule"),
  coding_system = "cow"
)

vdem_gw <- load_vdem_data(
  vars = c("v2x_polyarchy", "v2x_libdem"),
  coding_system = "gw"
)
}

}
