% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_rdata_from_extdata}
\alias{read_rdata_from_extdata}
\title{Load an RData file from package extdata}
\usage{
read_rdata_from_extdata(filename)
}
\arguments{
\item{filename}{Character. Name of the \code{.RData} file located in
\code{inst/extdata}.}
}
\value{
The single object stored in the \code{.RData} file.
}
\description{
Load an RData file from package extdata
}
\keyword{internal}
